feat: add alignment and offset properties to Wasm instructions for improved memory access

+++ WebAssemblySharp/Runtime/Interpreter/IWebAssemblyInterpreterMemoryArea.cs
@@ -8,5 +8,5 @@ namespace WebAssemblySharp.Runtime.Interpreter;
  */
 public interface IWebAssemblyInterpreterMemoryArea
 {
-    Span<byte> GetMemoryAccess(int p_Address, int p_Length);
+    Span<byte> GetMemoryAccess(long p_Address, int p_Length);
 }
